Use zone aware date formatter to render TimeSegment.
TimeSegment is used to render vertical time series at the left side in the day view.
It was not time zone aware. Now it is. Time is always rendered in devices default time zone.
So far, this is the only place a date is rendered in devices timezone.
Maybe we should consider to always display dates in the conferences timezone?

See also: #92

---

Fixes:

- TimeSegmentTest