Remove MyApp.roomList
This field was used to map from column index to room index.
But with the introduction of LectureListTransformer
(commit 08c0e2943424454025eb120f60c5ec78ac012fcc) the column
index is equal to the room index. Removing the mapping means
we can simply use the column index as the room index and
vice versa.