Restore the "roomIndex fix"
roomIndex is still used in a couple of places. We need to keep
this fix in place until we changed all UI code to not use
roomIndex at all.

This fixes a bug (the room name in the details screen does not
match the room name in the schedule screen) that was introduced
with the removal of the "roomIndex fixup" code in
1b3df482d589fa838f789db22ed26a0f888d1830.