Remove roomIndex rewriting from LectureListTransformer
The Lecture.roomIndex field as used by the UI layer had different semantics
than the "database value". That's why we had to rewrite the field before
handing it off to the UI layer. Now that the UI layer no longer accesses the
field we can get rid of the rewriting.