Consistently use "session" term in "network" module (file names).
Rename files:

- Lecture -> Session
- LectureTest -> SessionTest

Note that changes are intentionally scoped to the "network" module
as much as possible to keep this commit small and readable. Further,
files are just renamed without changing their content to being able
to follow the renaming in the commit history. Changes to the content
such as variable and function names are being made in a separate
subsequent commit.