Improve 'ScrollToColumn' Function
The Function is a setter for the 'activeColumnIndex' variable, which indicates the most left displayed column.
This commit improves on calculating the maximum value it can have.

Previously, this was done by dividing the total with of the ScrollView through the individual width of the columns,  calculating the total column amount.
Then, as the index starts at 0, 1 as subtracted. This didn't work for cases where more than one column was displayed, e.g. Tablets and Phones in Landscape mode.

To improve this, instead of dividing the ScrollView's width, the 'roomCount' variable was used, as it already provides that value. Instead of 1, the amount of Columns displayed was then subtracted to correct for different Amounts of Columns displayed