Improve 'onFling' Method
There was an unnecessary try/catch statement removed.

Also, the previous implementation split scrolling to the left and right into separate if-statements, although they were basically doing the same.
This was replaced by one if-statement, checking if the distance and velocity of the swipe met the requirements for a fling-scroll.
Also, instead of calculating the amount of columns to scroll as an unsigned value and adding or subtracting it depending on the swipe-direction, the amount is now directly calculated as a signed value.