Fix Overscroll Behavior in Android 12+
The error was caused by the OnTouchListener catching OnTouch Events even though the target columnIndex was outside the allowed range.

This was fixed by adding a check to the OnTouchListener to make sure the columnIndex is within the allowed range.
For this, the 'calculateOnTouchColumnIndex()' Method was changed to return '-1' if the columnIndex is outside the allowed range.

Also, there was a faulty JaCoCo test that should check for the columnIndex being to big in portrait mode but didn't produce an index outside of the allowed range.