Implement new notifications permission.
Since Android 13 there is a permission that has to be granted
in order to send notifications.

I implemented it so that when the permission is not granted,
the app won't let the user add alarms and instead show a Toast.
Also, when the app is started without notification permissions
and the user has alarms set, a Toast is displayed.