Create `InMemorySearchRepository` for tests
The tested behavior of `SearchHistoryManager` is changed to make it possible to simply use a `MutableStateFlow` for the `SearchRepository` implementation. The previous behavior (flow emits a list that is equal to the previous list after an item has been appended that already was in the list) did nothing useful – it would trigger an UI update even though nothing had changed.

The `SearchRepository` implementation in `AppRepository` is updated to exhibit the same behavior as `InMemorySearchRepository`.